@@ -14,28 +14,35 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.0.1",
-    "@angular/common": "~13.0.1",
-    "@angular/compiler": "~13.0.1",
-    "@angular/core": "~13.0.1",
-    "@angular/forms": "~13.0.1",
-    "@angular/localize": "~13.0.1",
-    "@angular/platform-browser": "~13.0.1",
-    "@angular/platform-browser-dynamic": "~13.0.1",
-    "@angular/router": "~13.0.1",
+    "@angular/animations": "^14.2.12",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/forms": "^14.2.12",
+    "@angular/localize": "^14.2.12",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
+    "@fortawesome/angular-fontawesome": "^0.10.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.36",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "@ngrx/effects": "^14.3.2",
+    "@ngrx/store": "^14.3.2",
+    "@ngrx/store-devtools": "^14.3.2",
     "@progress/kendo-angular-buttons": "^8.0.0",
     "@progress/kendo-angular-charts": "^7.1.1",
-    "@progress/kendo-angular-common": "^3.1.0",
-    "@progress/kendo-angular-dateinputs": "^7.1.0",
+    "@progress/kendo-angular-common": "^3.0.0",
+    "@progress/kendo-angular-dateinputs": "^7.0.0",
     "@progress/kendo-angular-dialog": "^7.1.2",
-    "@progress/kendo-angular-dropdowns": "^7.0.3",
+    "@progress/kendo-angular-dropdowns": "^7.0.0",
     "@progress/kendo-angular-editor": "^4.1.7",
-    "@progress/kendo-angular-excel-export": "^5.0.1",
+    "@progress/kendo-angular-excel-export": "^5.0.0",
     "@progress/kendo-angular-grid": "^7.2.4",
-    "@progress/kendo-angular-inputs": "^9.0.4",
-    "@progress/kendo-angular-intl": "^4.1.0",
+    "@progress/kendo-angular-icons": "^2.0.0",
+    "@progress/kendo-angular-inputs": "^9.0.0",
+    "@progress/kendo-angular-intl": "^4.0.0",
     "@progress/kendo-angular-l10n": "^4.0.0",
-    "@progress/kendo-angular-label": "^4.0.1",
+    "@progress/kendo-angular-label": "^4.0.0",
     "@progress/kendo-angular-layout": "^7.1.1",
     "@progress/kendo-angular-notification": "^4.0.0",
     "@progress/kendo-angular-pdf-export": "^4.0.0",
@@ -77,7 +86,6 @@
     "karma-jasmine-html-reporter": "^2.0.0",
     "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.4.4"
+    "typescript": "~4.8.4"
   }
 }
