@@ -2,18 +2,10 @@ import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
 import { NgModule, LOCALE_ID } from '@angular/core';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { HttpClientModule } from '@angular/common/http';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { AppComponent } from './app.component';
-import { RatingComponent } from './components/team/rating.component';
-import { DashboardComponent } from './components/dashboard/dashboard.component';
-import { CardComponent } from './components/planning/cards/card.component';
-import { PlanningComponent } from './components/planning/planning.component';
-import { ProfileComponent } from './components/profile/profile.component';
-import { InfoComponent } from './components/info/info.component';
-import { HeaderComponent } from './header/header.component';
-import { TeamComponent } from './components/team/team.component';
 import { CustomMessagesService } from './services/custom-messages.service';
 
 import { ExcelModule, GridModule, PDFModule } from '@progress/kendo-angular-grid';
@@ -31,33 +23,61 @@ import { DropDownsModule } from '@progress/kendo-angular-dropdowns';
 import { NotificationModule } from '@progress/kendo-angular-notification';
 import { MessageService } from '@progress/kendo-angular-l10n';
 
-const drawerRoutes = [
-    { path: '', component: TeamComponent },
-    { path: 'dashboard', component: DashboardComponent },
-    { path: 'planning', component: PlanningComponent },
-    { path: 'profile', component: ProfileComponent },
-    { path: 'info', component: InfoComponent }
-];
+
+
+// Import routing module
+import { AppRoutingModule } from './app-routing.module';
+
+// Import containers
+import {
+    DefaultFooterComponent,
+    DefaultHeaderComponent,
+    DefaultLayoutComponent,
+    StaticLayoutComponent,
+    StaticHeaderComponent,
+    StaticFooterComponent
+
+  } from './containers';
+
+const APP_CONTAINERS = [
+    DefaultFooterComponent,
+    DefaultHeaderComponent,
+    DefaultLayoutComponent,
+    StaticLayoutComponent,
+    StaticHeaderComponent,
+    StaticFooterComponent
+  ];
+
+import {  } from './containers/static-layout/static-layout.component';
+import {  } from './containers/static-layout/static-header/static-header.component';
+import {  } from './containers/static-layout/static-footer/static-footer.component';
 
 import 'hammerjs';
 
 import '@progress/kendo-angular-intl/locales/en/all';
 import '@progress/kendo-angular-intl/locales/es/all';
 import '@progress/kendo-angular-intl/locales/fr/all';
+import { StoreModule } from '@ngrx/store';
+import { reducers, effects } from './store/index';
+import { EffectsModule } from '@ngrx/effects';
+import { StoreDevtoolsModule } from '@ngrx/store-devtools';
+import { environment } from '../environments/environment.dev';
+import { AuthInterceptor } from './auth-interceptor';
+import { PerfectScrollbarModule } from 'ngx-perfect-scrollbar';
+import { IconsModule } from '@progress/kendo-angular-icons';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { TreeListModule } from '@progress/kendo-angular-treelist';
+
+
 
 @NgModule({
     declarations: [
         AppComponent,
-        RatingComponent,
-        DashboardComponent,
-        CardComponent,
-        PlanningComponent,
-        ProfileComponent,
-        HeaderComponent,
-        InfoComponent,
-        TeamComponent
+        ...APP_CONTAINERS,
+        StaticHeaderComponent
     ],
     imports: [
+        AppRoutingModule,
         BrowserModule,
         BrowserAnimationsModule,
         FormsModule,
@@ -77,12 +97,25 @@ import '@progress/kendo-angular-intl/locales/fr/all';
         DateInputsModule,
         InputsModule,
         DropDownsModule,
-        RouterModule.forRoot(drawerRoutes),
-        NotificationModule
+        PerfectScrollbarModule,
+        NotificationModule,
+        StoreModule.forRoot(reducers, {
+            runtimeChecks: {
+              strictActionImmutability: true,
+              strictStateImmutability: true,
+            }
+          }),
+          EffectsModule.forRoot(effects),
+          HttpClientModule,
+          StoreDevtoolsModule.instrument({ maxAge: 50, logOnly: environment.production }),
+          IconsModule,
+          FontAwesomeModule,
+          TreeListModule,
     ],
     providers: [
         { provide: MessageService, useClass: CustomMessagesService },
-        { provide: LOCALE_ID, useValue: 'en-US' }
+        { provide: LOCALE_ID, useValue: 'en-US' },
+        { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true }
     ],
     bootstrap: [AppComponent]
 })
