@@ -7,7 +7,7 @@ namespace mipBackend.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class UsuarioController: ControllerBase
+    public class UsuarioController : ControllerBase
     {
 
         private readonly IUsuarioRepository _repository;
@@ -23,12 +23,24 @@ namespace mipBackend.Controllers
         public async Task<ActionResult<IEnumerable<UsuarioResponseDto>>> GetUsuario()
         {
 
+
+
             var Usuarios = await _repository.GetAllUsuarios();
             return Ok(Usuarios);
 
         }
 
 
+        [HttpPost]
+        public async Task<ActionResult<IEnumerable<UsuarioRegistroResponseDto>>> Registrar
+            (
+                [FromBody] UsuarioRegistroRequestDto request
+            )
+        {
 
+            var usuario = await _repository.RegistroUsuario(request);
+
+            return Ok(usuario);
+        }
     }
 }
