@@ -5,13 +5,9 @@ namespace mipBackend.Data.Usuarios
     public interface IUsuarioRepository
     {
 
-        Task<UserAuthReponseDto> GetUsuario();
-
-        Task<UserAuthReponseDto> Login(UsuarioLoginRequestDto request);
-        
-        Task<UserAuthReponseDto> RegistroUsuario(UsuarioRegistroRequestDto request);
-
         Task<IEnumerable<UsuarioResponseDto>> GetAllUsuarios();
+
+        Task<IEnumerable<UsuarioRegistroResponseDto>> RegistroUsuario(UsuarioRegistroRequestDto request);
 
     }
 }
