@@ -6,6 +6,7 @@ using mipBackend.Models;
 using mipBackend.Token;
 using System.Net;
 using AutoMapper;
+using Microsoft.Data.SqlClient;
 
 namespace mipBackend.Data.Usuarios
 {
@@ -52,6 +53,45 @@ namespace mipBackend.Data.Usuarios
 
         public async Task<UserAuthReponseDto> GetUsuario()
         {
+
+            using (var db = _contexto)
+            {
+                var query = await (from usr in db.Users!
+                                   join urol in db.UserRoles! on usr.Id equals urol.UserId
+                                   join rol in db.Roles! on urol.RoleId equals rol.Id
+                                   join rol in db.Roles! on urol.RoleId equals rol.Id
+                                   where (niv.wkf03activo == 1 && wf.wkf01activo == 1 && tip.wkf02activo == 1)
+                                   orderby wf.wkf01llavepadre, wf.wkf01orden,
+                                   wf.wkf01prioridad
+
+                                   select new WorkflowResponseDto
+                                   {
+                                       wkf01llave = wf.wkf01llave,
+                                       wkf01nombre = wf.wkf01nombre,
+                                       wkf01descripcion = wf.wkf01descripcion,
+                                       wkf01llavepadre = wf.wkf01llavepadre,
+                                       wkf02llave = tip.wkf02llave,
+                                       wkf02nombre = tip.wkf02nombre,
+                                       wkf03llave = niv.wkf03llave,
+                                       wkf03nombre = niv.wkf03nombre,
+                                       wkf01esinicio = wf.wkf01esinicio,
+                                       wkf01orden = wf.wkf01orden,
+                                       wkf01prioridad = wf.wkf01prioridad,
+                                       wkf01activo = wf.wkf01activo,
+                                       wkf01directorio = wf.wkf01nombre,
+                                       wkf01url = wf.wkf01nombre,
+                                       wkf01iconourl = wf.wkf01nombre,
+                                       wkf01visiblemenu = wf.wkf01llave,
+                                       wkf01imagengrande = wf.wkf01nombre,
+                                       wkf01imagenpequena = wf.wkf01nombre,
+                                       wkf01estadoregistro = wf.wkf01nombre,
+
+                                   }).ToListAsync();
+
+                return _mapper.Map<IEnumerable<WorkflowResponseDto>>(query!);
+
+            }
+
             var usuario = await _userManager.FindByNameAsync(_usuarioSesion.ObtenerUsuarioSesion());
             if(usuario == null)
             {
@@ -94,9 +134,11 @@ namespace mipBackend.Data.Usuarios
 
         }
 
-        public async Task<UserAuthReponseDto> RegistroUsuario(UsuarioRegistroRequestDto request)
+        public async Task<IEnumerable<UsuarioRegistroResponseDto>> RegistroUsuario(UsuarioRegistroRequestDto request)
         {
-            var existeEmail = await _contexto.Users.Where(x => x.Email == request.Email).AnyAsync();
+
+            
+            var existeEmail = await _contexto.Users.Where(x => x.Email == request.UserName).AnyAsync();
 
             if (existeEmail)
             {
@@ -106,6 +148,17 @@ namespace mipBackend.Data.Usuarios
                 );
             }
 
+            var createdby = await _userManager.FindByNameAsync(_usuarioSesion.ObtenerUsuarioSesion());
+
+            if (createdby is null)
+            {
+                throw new MiddlewareException(
+                    HttpStatusCode.Unauthorized,
+                    new { mensaje = "El usuario no es valido para hacer esta insercion" }
+                    );
+            }
+
+            
             var existeUserName = await _contexto.Users.Where(x => x.UserName == request.UserName).AnyAsync();
 
             if (existeUserName)
@@ -116,24 +169,45 @@ namespace mipBackend.Data.Usuarios
                 );
             }
 
-            
-
             var usuario = new Usuario
             {
-                Email = request.Email,
+                Email = request.UserName,
                 UserName = request.UserName,
             };
+
 
             var resultado = await _userManager.CreateAsync(usuario!, request.Password!);
 
             if (resultado.Succeeded)
             {
 
-                return TransformerUserToUserDto(usuario);
+                
+
+                string sql = "EXEC pa_mipnet_usuarioDefault_i @NAME, @USERNAME, @PASSWORD, @USUARIO_ID ";
+
+                List<SqlParameter> parms = new List<SqlParameter>
+                {
+                    // Create parameter(s)    
+                    new SqlParameter { ParameterName = "@NAME", Value = request.Name },
+                    new SqlParameter { ParameterName = "@USERNAME", Value = request.UserName },
+                    new SqlParameter { ParameterName = "@PASSWORD", Value = request.Password },
+                    new SqlParameter { ParameterName = "@USUARIO_ID", Value = Guid.Parse(createdby.Id)}
+
+                };
+
+                var result = await _contexto.UsuarioRegistros!.FromSqlRaw<UsuarioRegistroResponseDto>(sql, parms.ToArray())
+                    .IgnoreQueryFilters()
+                    .ToListAsync();
+
+                return _mapper.Map<IEnumerable<UsuarioRegistroResponseDto>>(result);
 
             }
 
-            throw new Exception("EL email del usario no existe en la base de datos");
+           
+            throw new MiddlewareException(
+                   HttpStatusCode.Unauthorized,
+                   new { mensaje = resultado.Errors }
+               );
 
         }
 
