@@ -2,7 +2,7 @@
 {
     public class UsuarioRegistroRequestDto
     {
-        public string? Email { get; set; }
+        public string? Name { get; set; }
 
         public string? UserName { get; set; }
 
