@@ -5788,30 +5788,36 @@ namespace mipBackend.Migrations
                         .HasColumnName("fechaeliminacion");
 
                     b.HasKey("Prf03Llave")
-                        .HasName("PK_PRF03_PlantillaPerfil");
+                        .HasName("PK_PRF03_PlantillaFlujo");
 
                     b.ToTable("PRF03_Plantilla", (string)null);
                 });
 
-            modelBuilder.Entity("mipBackend.Models.Prf04PlantillaPerfil", b =>
+            modelBuilder.Entity("mipBackend.Models.Prf04PlantillaFlujo", b =>
                 {
+
+                    b.Property<int>("Prf04Llave")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
+                        .HasColumnName("PRF04_Llave");
+
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Prf04Llave"), 1L, 1);
+
                     b.Property<int>("Prf03Llave")
                         .HasColumnType("int")
                         .HasColumnName("PRF03_Llave");
 
-                    b.Property<int>("Wkf06Llave")
+                    b.Property<int>("Wkf01Llave")
                         .HasColumnType("int")
-                        .HasColumnName("WKF06_llave");
+                        .HasColumnName("WKF01_Llave");
 
-                    b.Property<int?>("Prf04Activo")
-                        .HasColumnType("int")
-                        .HasColumnName("PRF04_activo");
 
-                    b.HasKey("Prf03Llave", "Wkf06Llave");
 
-                    b.HasIndex("Wkf06Llave");
+                    b.HasKey("Prf04Llave");
 
-                    b.ToTable("PRF04_plantillaPerfil", (string)null);
+                    b.HasIndex("Wkf01Llave");
+
+                    b.ToTable("PRF04_PlantillaFlujo", (string)null);
                 });
 
             modelBuilder.Entity("mipBackend.Models.Prf05TipoAsignacionUsuario", b =>
@@ -6789,7 +6795,7 @@ namespace mipBackend.Migrations
                         .ValueGeneratedOnAdd()
                         .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Value"), 1L, 1);
+                    
 
                     b.Property<string>("Description")
                         .HasColumnType("nvarchar(max)");
@@ -7890,7 +7896,7 @@ namespace mipBackend.Migrations
 
                     b.Property<int?>("Wkf01Llave")
                         .HasColumnType("int")
-                        .HasColumnName("WKF01_llave");
+                        .HasColumnName("WKF01_Llave");
 
                     b.Property<int?>("Wkf06Activo")
                         .HasColumnType("int")
@@ -9100,23 +9106,23 @@ namespace mipBackend.Migrations
                     b.Navigation("Secu02LlaveNavigation");
                 });
 
-            modelBuilder.Entity("mipBackend.Models.Prf04PlantillaPerfil", b =>
+            modelBuilder.Entity("mipBackend.Models.Prf04PlantillaFlujo", b =>
                 {
                     b.HasOne("mipBackend.Models.Prf03Plantilla", "Prf03LlaveNavigation")
-                        .WithMany("Prf04PlantillaPerfils")
+                        .WithMany("Prf03Plantillas")
                         .HasForeignKey("Prf03Llave")
                         .IsRequired()
-                        .HasConstraintName("FK_PRF04_plantillaPerfil_PRF03_Plantilla");
+                        .HasConstraintName("FK_PRF04_PlantillaFlujo_PRF03_Plantilla");
 
-                    b.HasOne("mipBackend.Models.Wkf06Perfil", "Wkf06LlaveNavigation")
-                        .WithMany("Prf04PlantillaPerfils")
-                        .HasForeignKey("Wkf06Llave")
+                    b.HasOne("mipBackend.Models.Wkf01Workflow", "Wkf06LlaveNavigation")
+                        .WithMany("Prf04PlantillaFlujos")
+                        .HasForeignKey("Wkf01Llave")
                         .IsRequired()
-                        .HasConstraintName("FK_PRF04_plantillaPerfil_WKF06_Perfiles");
+                        .HasConstraintName("FK_PRF04_PlantillaFlujo_WKF01_Flujo");
 
                     b.Navigation("Prf03LlaveNavigation");
 
-                    b.Navigation("Wkf06LlaveNavigation");
+                    b.Navigation("Wkf01LlaveNavigation");
                 });
 
             modelBuilder.Entity("mipBackend.Models.Prf06PermisosUsuario", b =>
@@ -9772,7 +9778,7 @@ namespace mipBackend.Migrations
 
             modelBuilder.Entity("mipBackend.Models.Prf03Plantilla", b =>
                 {
-                    b.Navigation("Prf04PlantillaPerfils");
+                    b.Navigation("Prf04PlantillaFlujos");
                 });
 
             modelBuilder.Entity("mipBackend.Models.Prf05TipoAsignacionUsuario", b =>
@@ -9914,7 +9920,7 @@ namespace mipBackend.Migrations
 
             modelBuilder.Entity("mipBackend.Models.Wkf06Perfil", b =>
                 {
-                    b.Navigation("Prf04PlantillaPerfils");
+                    b.Navigation("Prf04PlantillaFlujos");
 
                     b.Navigation("Prf06PermisosUsuarios");
 
