@@ -13,7 +13,7 @@ namespace mipBackend.Migrations
                 columns: table => new
                 {
                     Value = table.Column<int>(type: "int", nullable: false)
-                        .Annotation("SqlServer:Identity", "1, 1"),
+                        ,
                     Description = table.Column<string>(type: "nvarchar(max)", nullable: true)
                 },
                 constraints: table =>
