@@ -5,7 +5,7 @@ namespace mipBackend.Models
     public class setSelect
     {
         [Key]
-        public int Value { get; set; }
+        public int? Value { get; set; }
 
         public string? Description { get; set; }
     }
