@@ -38,7 +38,7 @@ const routes: Routes = [
         path: 'dashboard',
         loadChildren: () =>
           import('./views/dashboard/dashboard.module').then((m) => m.DashboardModule),
-          canActivate: [AuthGuard]
+
       },
       {
         path: '',
